requirements.txt
# Core Framework
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
streamlit-aggrid>=0.3.4
streamlit-elements>=0.1.0
streamlit-plotly-events>=0.0.6

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
polars>=0.19.0

# Visualization
plotly>=5.17.0
altair>=5.1.0
seaborn>=0.12.0
matplotlib>=3.7.0

# Financial Analysis
quantlib>=1.32
factor-analyzer>=0.4.0

# Machine Learning
scikit-learn>=1.3.0
torch>=2.0.0
optuna>=3.4.0

# Backtesting & Optimization
backtrader>=1.9.78
vectorbt>=0.25.0
cvxpy>=1.4.0
scipy>=1.11.0

# Utilities
openpyxl>=3.1.0
xlrd>=2.0.1
networkx>=3.1
tqdm>=4.65.0

# Yahoo Finance API
yfinance>=0.2.28

# Technical Analysis
ta>=0.10.0

# Environment Management
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0 
